CSAcademy, CodeForces and TopCoder

+++ Voltage Keepsake/src.cpp
@@ -2,19 +2,19 @@
 
 typedef long long ll;
 
-const int maxn = 1e5 + 1e2;
+const ll maxn = 1e5 + 1e2;
 const double eps = 0.000000001;
-int drain[maxn], level[maxn];
+ll drain[maxn], level[maxn];
 
 int main(){
 	std::ios::sync_with_stdio(false);
 
-	int n, charge;
+	ll n, charge;
 	std::cin >> n >> charge;
 
-	int per_second_extra_charge = charge;
+	ll per_second_extra_charge = charge;
 
-	for(int i = 0; i < n; i++){
+	for(ll i = 0; i < n; i++){
 		std::cin >> drain[i] >> level[i];
 		per_second_extra_charge -= drain[i];
 	}
@@ -26,11 +26,11 @@ int main(){
 
 	double l = 0, r = 1e18;
 
-	for(int i = 0; i < 150; i++){
+	for(ll i = 0; i < 150; i++){
 		double mid = (l + r) / 2;
 
 		double needed = 0.0;
-		for(int j = 0; j < n; j++){
+		for(ll j = 0; j < n; j++){
 			needed += std::max(0.0, ((double) drain[j]) * mid - level[j]);
 		}
